new_requirements.txt
flask
yfinance
mplfinance
python-dotenv
line-bot-sdk==3.*
